스캐줄 앱 과제 발표 및 피드백 수정

+++ scheduleapp/app/src/main/java/com/bitcamp/myapp/vo/Schedule.java
@@ -3,6 +3,6 @@ package com.bitcamp.myapp.vo;
 public class Schedule {
   public int no;
   public String scheduleTitle;
-  public Double startTime;
-  public Double endTime;
+  public long startTime;
+  public long endTime;
 }
